% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcsinbacktran.R
\name{arcsinbacktran}
\alias{arcsinbacktran}
\title{Arcsine back-transformation}
\usage{
arcsinbacktran(x)
}
\arguments{
\item{x}{Arcsine-transformed proportion(s) to be back-transformed back to proportions.}
}
\value{
numeric vector with the original proportions
}
\description{
Given a set of arcsine-transformated proportions X, back-transforms them back to "raw" proportions; that is, undoes the arcsine transformation.
}
\examples{
mydata <- data.frame(Accuracy.Proportion = c(0.5,0.75,0.99,0, 0.2))
mydata$Accuracy.Arcsin <- arcsintran(mydata$Accuracy.Proportion)
mydata$Accuracy.Backtran <- arcsintran(mydata$Accuracy.Arcsin)
mydata # same as original proportions
}
\seealso{
\code{\link{arcsintran}} for the arcsine transformation
}
