% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contr.helmert.unweighted.R
\name{contr.helmert.unweighted}
\alias{contr.helmert.unweighted}
\title{Unweighted Helmert Contrast Matrices}
\usage{
contr.helmert.unweighted(x, reference.levels = levels[-length(levels)],
  n = NULL)
}
\arguments{
\item{x}{a factor variable (or variable that can be coerced to a factor) for
which contrasts should be calculated.}

\item{reference.levels}{vector specifying, in order, the category treated as
the reference level (i.e., assigned the next negative value) in each
successive contrast.}

\item{n}{a vector of levels for a factor, or the number of levels, which can
be provided instead of \code{x}.}
}
\value{
A matrix with \code{n} rows and \code{k} columns, where \code{n} is
the number of levels and \code{k=n-1}.
}
\description{
Returns a matrix of Helmert contrasts, scaled so that the resulting contrast
estimates (in an ANOVA or regression model) correspond to the difference
between the levels (categories) being compared. The contrasts may be computed
either based on a numerical number of levels or a vector of data.
}
\details{
Helmert contrasts compare the second level with the first, the third with the
average of the first two, and so on. As with other contrasts, they are
orthogonal to each other and to the intercept.

When the levels differ in frequency, \emph{unweighted} coding is appropriate
if the differences in frequency in the sample are merely \emph{incidental}
(e.g., both conditions were intended to be presented equally frequently, but
by chance there are more observations from one condition than another).  (If
the differences in frequency instead represent genuine differences in the
population, weighted coding may be more appropriate.)

If all of the factor levels are equally common, there is no difference
between unweighted and weighted coding.
}
\examples{
contr.helmert.unweighted(n=4)

contr.helmert.unweighted(n=c('Active','Passive'), reference.levels=1)

cuedata <- as.factor(c('ValidCue', 'ValidCue',
   'InvalidCue', 'NoCue', 'InvalidCue', 'NoCue'))
contr.helmert.unweighted(x=cuedata)
contr.helmert.unweighted(x=cuedata,
                         reference.levels=c('ValidCue','InvalidCue'))
}
\references{
Cohen, J., Cohen, P., West, S.G., & Aiken, L.S. (2002). Categorical or
nominal independent variables. In \emph{Applied multiple regression/
correlation analysis for the behavioral sciences} (3rd ed., pp. 302-353).
Mahwah, NJ: Lawrence Erlbaum Associates.
}
\seealso{
\code{\link{contr.helmert.weighted}} for \emph{weighted} Helmert
contrasts, and \code{\link{contrasts}} and \code{\link{contr.helmert}}.
}
