% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.by.group.R
\name{cor.by.group}
\alias{cor.by.group}
\title{Correlation Within Groups}
\usage{
cor.by.group(x, y, INDEX, ...)
}
\arguments{
\item{x}{one of the numeric vectors to be correlated.}

\item{y}{the other numeric vector to be correlated.}

\item{INDEX}{list of one or more grouping variables, typically factors, each
of the same length as \code{x}.}

\item{...}{arguments to \code{\link{cor}}.}
}
\description{
Computes the correlation between \code{x} and \code{y} separately within each
of several groups defined by all of the combinations of the factor variables
contained within \code{by}. For example, correlate high school GPA with
college GPA separately within each of several universities, or correlate
working memory score with response time separately in each cell of a 2 x 2
factorial experiment.
}
\examples{
data(grouseticks, package='lme4')
cor.by.group(grouseticks$HEIGHT, grouseticks$TICKS, INDEX=grouseticks$YEAR)
# Correlation of HEIGHTS and TICKS separately for each YEAR

data(Arabidopsis, package='lme4')
cor.by.group(Arabidopsis$nutrient, Arabidopsis$total.fruits,
   INDEX=list(Arabidopsis$reg, Arabidopsis$status))
# Correlation of nutrient and total.fruits for each combination of
# reg x status
}
\seealso{
\code{\link{cor}} for correlations without division into groups.
}
