% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.as.numeric.R
\name{factor.as.numeric}
\alias{factor.as.numeric}
\title{Coerce Factor Labels to Numeric}
\usage{
factor.as.numeric(x)
}
\arguments{
\item{x}{factor variable whose labels should be coerced to numeric.}
}
\value{
numeric vector.
}
\description{
Coerces a factor to a numeric vector based on the labels for the factor 
levels (rather than the integer codes assigned to them); e.g., factor level
"3.5" becomes a value of 3.5.  Labels that cannot be coerced to numeric are
treated as NA.
}
\examples{
my.factor <- as.factor(c(1,2.5,10))
factor.as.numeric(my.factor)
# compare to:
as.numeric(my.factor)
}
