% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fence.by.group.R
\name{fence.by.group}
\alias{fence.by.group}
\title{Fence Values Based on Group Means and Standard Deviations}
\usage{
fence.by.group(x, INDEX, sds = 3, print = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{INDEX}{list of one or more grouping variables, typically factors, each
of the same length as \code{x}.}

\item{sds}{number of standard deviations away from the mean at which
values should be replaced.}

\item{print}{a logical value indicating whether to display the number and
percentage of replaced elements.}

\item{na.rm}{a logical value indicating whether NA values should be ignored
when computing the mean and standard deviation.}
}
\value{
vector in which values more than the specified number of standard
deviations from each cell mean have been replaced with the boundary value.
}
\description{
Identifies values of x that are a certain number of standard deviations from
the mean within a particular set of grouping variables and replaces them with
that boundary value (e.g., values more than 3 standard deviations from the
mean are replaced with the mean +/= 3 standard deviations).  For example,
fence response times that are more than 2.5 standard deviations from the mean
within each cell of a PrimeType x WordFrequency factorial experimental
design.
}
\examples{
data(sleepstudy, package='lme4')
sleepstudy$Reaction.Fenced <- fence.by.group(sleepstudy$Reaction,
  sleepstudy$Subject, sds=3) # fence RTs 3 std devs from subject mean

data(VerbAgg, package='lme4')
VerbAgg$Anger.Fenced <- fence.by.group(VerbAgg$Anger,
  list(VerbAgg$btype, VerbAgg$situ), sds=3)
# fence 3 std devs from behavior type x situation type cell mean
}
\seealso{
\code{\link{flag.by.group}} to identify such observations without
replacing them.

\code{\link{trim.by.group}} to treat such observations as missing
data rather than replacing them with a new value.

\code{\link{fence}} to fence values a certain number of standard
deviations away from the overall mean rather than a group mean.
}

