% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fence.R
\name{fence}
\alias{fence}
\title{Fence Values Based on Mean and Standard Deviation}
\usage{
fence(x, sds = 3, print = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{vector of values to which the fencing procedure should be applied.}

\item{sds}{number of standard deviations away from the mean at which values
should be replaced.}

\item{print}{a logical value indicating whether to display the number and
percentage of replaced values.}

\item{na.rm}{a logical value indicating whether NA values should be ignored
when computing the mean and standard deviation.}
}
\value{
vector in which values more than the specified number of standard
deviations from the mean have been replaced with the boundary value.
}
\description{
Identifies values of x that are a certain number of standard deviations from
the mean and replaces them with that boundary value (e.g., values more than
3 standard deviations from the mean are replaced with the mean +/= 3 standard
deviations).
}
\details{
This function is provided because this is a standard data processing
procedure in the psycholinguistic literature and not as an endorsement or
rejection of this or any other procedure.
}
\examples{
my.data <- data.frame(RT=rnorm(300, mean=789, sd=150))
my.data$FencedRT <- fence(my.data$RT, sds=2.5)
}
\seealso{
\code{\link{flag.by.sds}} to identify such observations without
replacing them.

\code{\link{trim}} to treat such observations as missing data rather
than replacing them with a new value.

\code{\link{fence.by.group}} to fence values a certain number of
standard deviations away from a group mean (e.g., the mean within a condition
or the mean for a research participant) rather than the overall mean.
}

