% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.by.sds.R
\name{flag.by.sds}
\alias{flag.by.sds}
\title{Identify Extreme Values Based on Mean and Standard Deviation}
\usage{
flag.by.sds(x, sds = 3, print = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{vector of values.}

\item{sds}{number of standard deviations away from the mean at which
"extreme" values should be identified.}

\item{print}{a logical value indicating whether to display the number and
percentage of identified elements.}

\item{na.rm}{a logical value indicating whether NA values should be ignored
when computing the mean and standard deviation.}
}
\value{
logical vector indicating whether each element of x is more than
the specificed number of deviations away from the mean.
}
\description{
Identifies which elements of x are a certain number of standard deviations
from the mean (e.g., values more than 3 standard deviations from the mean).
}
\examples{
my.data <- data.frame(RT=rnorm(300, mean=789, sd=150))
my.data$OutlyingRT <- flag.by.sds(my.data$RT, sds=2)
}
\seealso{
\code{\link{trim}} to treat such elements as missing data.

\code{\link{fence}} to \emph{replace} such elements with the
threshold value.

\code{\link{flag.by.group}} to identify values a certain number of
standard deviations away from a group mean (e.g., the mean within a condition
or the mean for a research participant) rather than the overall mean.
}
