% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.compare.R
\name{lmer.compare}
\alias{lmer.compare}
\title{Compare Multiple Mixed-Effects Models on AIC and BIC}
\usage{
lmer.compare(
  listofmodels,
  show.formulae = TRUE,
  summarize.aic = FALSE,
  summarize.bic = FALSE
)
}
\arguments{
\item{listofmodels}{a list of \code{merMod} objects.}

\item{show.formulae}{a logical value whether the model formula should be
displayed for each model.}

\item{summarize.aic}{a logical value indicating whether the model summary
should be displayed for the model that is the best-fitting according to
AIC.}

\item{summarize.bic}{a logical value indicating whether the model summary
should be displayed for the model that is the best-fitting according to
BIC.  Note that if \code{summarize.aic} and \code{summarize.bic} are both
\code{TRUE} but the same model wins according to both criteria, only one
summary will be displayed.}
}
\description{
Compares a set of mixed-effects models on their AIC and BIC values,
highlighting the best model according to each of these criteria.
}
\details{
Log likelihood and deviance are \emph{not} reported because likelihood-ratio
tests are only applicable for NESTED models, and the models may or may not
be nested.
}
