% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.rsquared.R
\name{lmer.rsquared}
\alias{lmer.rsquared}
\title{\eqn{{R}^2}{R^2} for Linear Mixed Effects Models}
\usage{
lmer.rsquared(model)
}
\arguments{
\item{model}{a model of class \code{merMod} for which the \eqn{{R}^2}{R^2}
value should be obtained.}
}
\value{
\eqn{{R}^2}{R^2} value between 0 and 1.
}
\description{
Returns the \eqn{{R}^2}{R^2} value for a linear mixed effect model.
}
\details{
In a mixed effects model, much of the variance will be "explained" in a
theoretically uninformative way simply by the random effects (e.g., "some
people are different from others").  To the extent that an \eqn{{R}^2}{R^2}
value is useful for these models, it will be useful when comparing the
\eqn{{R}^2}{R^2} of a model that incorporates some fixed effect(s) of
interest to that of a model that does not.
}
\references{
Baayen, R.H. (2008). \emph{Analyzing linguistic data: A practical
introduction to statistics}. Cambridge, UK: Cambridge University Press.

Baayen, R.H., & Milin, P. (2010). \emph{Analyzing reacion times},
\emph{3}, 12-28.
}
