% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logittoprob.R
\name{logittoprob}
\alias{logittoprob}
\title{Logit-to-Proportion Transformation}
\usage{
logittoprob(x)
}
\arguments{
\item{x}{logit value to transformed into a proportion}
}
\description{
Back-transforms logits to proportions.  The logit
represents the log odds of a "hit" or a 1--that is, log(p(hit)/(1-p(hit))--
and is a better transformation to proportions for approximating the normal
distribution than is the arcsine transformation.
}
\examples{
logittoprob(1.37)
logittoprob(logit(0.25))

}
\seealso{
\code{\link{logit}} to transform proportions into logits.
}
