% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.zero.R
\name{pad.zero}
\alias{pad.zero}
\title{Pad String With Leading Zeros}
\usage{
pad.zero(x, width)
}
\arguments{
\item{x}{a numeric vector or vector that can be coerced to numeric.}

\item{width}{minimum width desired for each string.}
}
\value{
character vector of digit strings in which 0s have been prepended to
each string to make it at least \code{width} characters long.
}
\description{
Pads each element of a numeric vector, or a vector that can be coerced to
numeric, so that it is a string of a minimum \code{width} by adding leading
0s.
}
\details{
Non-integers are truncated towards 0.

For negative integers, the negative sign is counted towards the string
length.

Integers that are already as long or longer than \code{width} are unaffected.
}
\examples{
pad.zero(c('3','80','155'), width=2)
pad.zero(20, width=3)
}
