% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probdifftologit.R
\name{probdifftologit}
\alias{probdifftologit}
\title{Proportion Difference to Logit}
\usage{
probdifftologit(x, intercept = 0.5)
}
\arguments{
\item{x}{A difference of proportions.}

\item{intercept}{The intercept for the model.}
}
\description{
Converts a difference in proportions to a difference of log odds (logits).
Note that the relationship of a difference in proportion to a difference in
logits depends on the positioning on the overall scale of propotions; thus,
this function also requires specifying an intercept term from the proportions.
}
\examples{
probdifftologit(x=.20, intercept=.50)
logit(.60) - logit (.40)
probdifftologit(x=.20, intercept=.75)
logit(.85) - logit (.65)

}
\seealso{
\code{\link{logit}} to transform individual probabilities into logits.
}
