% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stderror.R
\name{stderror}
\alias{stderror}
\title{Standard error}
\usage{
stderror(x, bessel = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{value(s) for which the standard error of the mean should be computed}

\item{bessel}{a logical value indicating whether Bessel's correction should
be applied to obtain an unbiased estimate of the variance of the sampling
distribution.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}
}
\value{
standard error of the mean of x
}
\description{
Calculates the standard error of the mean of x with or without applying the
Bessel correction (dividing by \emph{N}-1 rather than \emph{N}).  Default
behavior is to apply the correction.
}
\details{
Note that Bessel's correction provides an unbiased estimate of the
\emph{variance} of the sampling distribution, but the estimate of the
\emph{standard deviation} is still biased.
}
\examples{
stderror(c(150, 115, 130, 170, 190))
}
\references{
Bessel's correction. (2015, Aug 15). Retrieved from
http://en.wikipedia.org/wiki/Bessel's_correction

Hays, W.L. (1994). \emph{Statistics} (5th ed.). Belmont, CA:
Wadsworth.
}
