% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryCI.R
\name{summaryCI}
\alias{summaryCI}
\title{Linear Mixed Effect Model Summary with Confidence Intervals}
\usage{
summaryCI(model, confidence = 0.95)
}
\arguments{
\item{model}{\code{lmer} model to be summarized.}

\item{confidence}{desired level of confidence for the confidence interval.}
}
\value{
model summary including confidence intervals.
}
\description{
For a \code{lmer} model fit, displays a model summary with confidence
intervals. Confidence intervals are obtained by comparing the
\emph{t}-statistic to the standard normal distribution, which is appropriate
for models with a large number of degrees of freedom.
}
\examples{
model1 <- lmer(y ~ service + (1|s) + (1|d), data=InstEval)
summaryCI(model1)
}
\references{
Baayen, R.H. (2008). \emph{Analyzing linguistic data: A practical
introduction to statistics}. Cambridge, UK: Cambridge University Press.
}
