% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryExp.R
\name{summaryExp}
\alias{summaryExp}
\title{Generalized Linear Mixed Effect Model Summary with Exponeniated
Coefficients}
\usage{
summaryExp(model, confidence = 0.95)
}
\arguments{
\item{model}{\code{glmer} model to be summarized.}

\item{confidence}{desired level of confidence for the confidence interval.}
}
\value{
model summary including exponentiated fixed effect estimates and
confidence intervals.
}
\description{
For a \code{glmer} model fit using the logit or log link function, displays
a model summary with the exponentiation function applied to the fixed effect
point estimates. Thus, estimates of effects on the \emph{log odds} in a
logistic model become estimates of effects on the \emph{odds}, which are
usually more interpretable.  Similarly, estimates of effects on the \emph{log
counts} in a Poisson regression become estimate of effcts on the
\emph{counts} to facilitate interpretation.
}
\details{
Note that once the coefficients have been exponeniated, the estimates now
represent \emph{multiplicative} effects.  For instance, a variable associated
with an odds ratio of 1.25 in a logistic regression means the variable is
associated with 1.25 TIMES the odds ofsuccess.  (This is different from
regular linear regression, where the effects are additive.)

Consequently, a factor that has no effect on the response variable would have
an estimate of 1 rather than 0.  (Multiplying something by 1 leaves it
unchanged, of course.)

95% confidence intervals for the estimates are also displayed.  Because the
effects are multiplicative, the CIs will NOT be symmetric about the point
estimate on the original scale.  Thus, it would be preferable to report the
lower- and upper-bound of the CI (which is the current APA style for
confidence intervals anyway) rather than its width.
}
\examples{
data(VerbAgg, package='lme4')
model1 <- glmer(r2 ~ Gender + btype+ (1|id) + (1|item),
  family=binomial, data=VerbAgg)
summaryExp(model1)
}

