% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim Values Based on Mean and Standard Deviation}
\usage{
trim(x, sds = 3, print = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{vector of values to which the trimming procedure should be applied.}

\item{sds}{number of standard deviations away from the mean at which values
should be considered missing.}

\item{print}{a logical value indicating whether to display the number and
percentage of extreme values.}

\item{na.rm}{a logical value indicating whether \emph{existing} NA values
should be ignored when computing the mean and standard deviation.}
}
\value{
vector in which values more than the specified number of standard
deviations from the mean have been replaced with \code{NA}.
}
\description{
Identifies values of x that are a certain number of standard deviations from
the mean and treats them as missing (\code{NA}) values (e.g., values more
than 3 standard deviations from the mean are considered missing) so that they
can be easily deleted.
}
\details{
This function is provided because this is a standard data processing
procedure in the psycholinguistic literature and not as an endorsement or
rejection of this or any other procedure.
}
\examples{
my.data <- data.frame(RT=rnorm(300, mean=789, sd=150))
my.data$TrimmedRT <- trim(my.data$RT, sds=2.5)
}
\seealso{
\code{\link{flag.by.sds}} to identify such values without deleting
them.

\code{\link{fence}} to replace such observations with the boundary
value.

\code{\link{trim.by.group}} to trim values a certain number of
standard deviations away from a group mean (e.g., the mean within a condition
or the mean for a research participant) rather than the overall mean.
}
