% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.by.group.R
\name{zscore.by.group}
\alias{zscore.by.group}
\title{z-Score Within Each of Several Cells}
\usage{
zscore.by.group(x, INDEX)
}
\arguments{
\item{x}{a numeric vector}

\item{INDEX}{list of one or more grouping variables, typically factors, each 
of the same length as x}
}
\value{
numeric vector with \code{x} z-scored separately within each level
of \code{INDEX}
}
\description{
z-scores \code{x} within each of one or more cells.  For instance, z-score
Likert scale ratings within each participant, or z-score response times
within each participant x condition pairing.
}
\examples{
my.data <- data.frame(Subject=c(rep('S1',6), rep('S2',6)),
                      Session=c(rep('Day1',3), rep('Day2',3),
                                rep('Day1',3), rep('Day2',3)),
                      Rating=c(10,30,20,20,40,30,15,35,25,25,45,35))
# z-score within each subject:
my.data$Rating.Subject.z <- zscore.by.group(my.data$Rating, my.data$Subject)
# z-score within each subject x session pairing:
my.data$Rating.SubjectSession.z <- zscore.by.group(my.data$Rating,
list(my.data$Subject, my.data$Session))
}
\seealso{
\code{\link{center.by.group}} to center within groups without
scaling by the standard deviation.
}
